🐛 fix(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics ✨ feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port The port variable is now named PORT, which improves consistency with the naming conventions as PORT is a constant. Support for an environment variable allows the application to be more flexible as it can now run on any available port specified via the process.env.PORT environment variable.
🚚 chore(ConfigureServices.cs): remove unused code and add missing dependencies
The unused code related to adding FluentValidationAutoValidation and HttpClient has been removed. The missing dependencies IProxyFactory and ICurrentCancellationTokenService have been added to the ConfigureServices class.

✨ feat(ICancellationTokenService.cs): add interface for CancellationTokenService
An interface ICurrentCancellationTokenService has been added to the Common.Interfaces namespace. This interface defines a CancellationToken property.

✨ feat(IProxyFactory.cs): add interface for ProxyFactory
An interface IProxyFactory has been added to the Common.Interfaces namespace. This interface defines a GetProxy method that returns a tuple containing a Stream, a processId, and a ProxyStreamError.

✨ feat(HttpClientExtensions.cs): remove unused code
The byteRangeStart

🔨 refactor(StreamManager.cs): add proxyFactory parameter to StreamManager constructor to improve extensibility and flexibility
The StreamManager class now accepts an additional parameter, `IProxyFactory`, in its constructor. This change was made to improve the extensibility and flexibility of the StreamManager class. The `IProxyFactory` parameter allows for the injection of a proxy factory implementation, which can be used by the StreamManager to create proxy objects as needed. This change ensures that the StreamManager class is more modular and can easily support different proxy implementations in the future.
