+++ StreamMasterApplication/Common/Interfaces/ICancellationTokenService.cs
@@ -1,0 +1,6 @@
+﻿namespace StreamMasterApplication.Common.Interfaces;
+
+public class ICurrentCancellationTokenService
+{
+    public CancellationToken CancellationToken { get; set; }
+}
