+++ StreamMasterApplication/Common/Interfaces/IProxyFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace StreamMasterApplication.Common.Interfaces
+{
+    public interface IProxyFactory
+    {
+        Task<(Stream? stream, int processId, ProxyStreamError? error)> GetProxy(string streamUrl);
+    }
+}
