+++ StreamMasterInfrastructure/Common/HttpClientExtensions.cs
@@ -7,8 +7,6 @@ public static class HttpClientExtensions
     public static async Task<HttpResponseMessage?> GetWithRedirectAsync(
         this HttpClient httpClient,
         string requestUri,
-        //long byteRangeStart = 0,
-        //long byteRangeLength = -1,
         int maxRetries = 5,
         CancellationToken cancellationToken = default)
     {
@@ -18,14 +16,8 @@ public static class HttpClientExtensions
 
         while (retryCount <= maxRetries)
         {
-            //if (byteRangeLength != -1)
-            //{
-            //    httpClient.DefaultRequestHeaders.Range = new RangeHeaderValue(byteRangeStart, byteRangeStart + byteRangeLength);
-            //}
-            //else
-            //{
-            //    httpClient.DefaultRequestHeaders.Range = null;
-            //}
+            httpClient.DefaultRequestHeaders.Range = null;
+
             try
             {
                 response = await httpClient.GetAsync(redirectUrl, HttpCompletionOption.ResponseHeadersRead, cancellationToken).ConfigureAwait(false); ;
